moved fundtions to  utilsCartContext functions
